hypnosis-backend/requirements.txt
virtualenv==13.1.0
djangorestframework==3.0.5
django-rest-framework-mongoengine==2.0.2
pymongo==2.8
mongoengine==0.9.0
Django==1.8.3
gunicorn==19.3.0
markdown
django-filter
jwt==0.3.2
django-rest-swagger
bson==0.3.3
